added some warnings to ignor List.
C4996 says that there are some functions are unsafe like "sprintf" and that one should use "sprintf_s" instead. since I dont know about an Linux equivalent and the current functions are working i put this message to ignore list.

the other since on the rest is just putting on both project files (vs2010 vs2012) the same things on ignore.