added functionality:
now one can send the event number to the ROOTProducer. this works as a crosscheck and the Root producer will warn you if there is a mismatch. The event number gets also written to the event as a tag this gives you offline the Possibility to recover any Event mismatch