bug fix: TLUController was crashing when one was using the flag -n which is suppressing the readout of the timestamps.
The new feature, which gives the possibility to read out the information which PMT has fired, added one more input parameter to the Constructor of TLUEntry.
this parameter was only set if the time stamp where read out otherwise it would stay a nullptr. the bugfix now checks if the trigger information was set or not. if not initializes it with zero