the event packet gets now processed correctly I am not sure if it should stay in but at least it works now with all the rest of the mechanism
there was a memory leak in event packet. it was holding the event with a raw pointer but it was never deleting it. this was fine for the case in which it was sending the data to disk because there the somebody else had the ownership but for the deserializing this class was the only thing that was having this pointer. now it is done with a shared pointer.