Fix problem with event ID mixing in case of running without a TLU.
Running with a TLU the data collector re-syncs its own event ID to
the one from the TLU. This means, BORE has event ID 0, but also the
first data event from gets ID 0 since this is the first ID sent by
the producers.
In case eudaq is ran without a TLU there is no resync happening and
the DataCollector's internal event ID is used - but this has been
increased after the BORE - leading to a mismatch of the ID of the
DetectorEvent with respect to the contained RawEvents. Now, the
DataCollector's counter is not incremented after the BORE fixing this.

This also fixes the OnlineMon problem "Plane Mismatch" when running
without TLU, since now an event 0 exists and is delivered.