CMSPixel: change behavior when reading multiple I2C devices
if no "i2c" parameter is given, the address defaults to "0" and
the config files are read as written in the config file. If "i2c"
is given, the file path is appended with the "_Cx" pattern, where
x is the i2c address. This works now for both single chips and
multiple ROCs.