BugFix:
when running 2 producer from the same function the logcollector was crashing.

the LogSender is a singleton class and each producer tries to connect the LogSender to the logcollector. in some cases this lead to a crash of the logcollector.
now the LogSender does not try to reconnect and now it works. both producer use now the same log class