Added comment on the producer ready mechanism:
This safety mechanism is needed for running with automatic configuration
file changes. It ensures, that the producers finished their
configuration, before the next run is started. If that is covered by the
control FSM in a later version of the software, these variables and the
respective code need to be removed again.