First commit of gui modification branch. This commit has changed the following files within the EUDAQ structure. Changes to be committed: 	modified:   gui/include/RunControlModel.hh 	modified:   gui/include/euRun.hh 	modified:   gui/src/RunControlModel.cc 	modified:   gui/src/euRun.cc
The changes include the addition of a function in the RunControlModel class
called CheckConfigured, which checkes each connection and returns true if and
only if they are all configured.

In the euRun class the startbtn clicked watcher function has been modified so
that the data collection will not start unless all connections are configured.

Minor other debug outputs has been added, as well as comments to some of the
previously existing functions.