Changes to be committed: 	modified:   gui/include/euRun.hh 	modified:   gui/src/euRun.cc This commit takes care of an issue involving the user's ability to cofigure a producer before a datacollector was present, thus inducing a segmentation fault that crashed the program. The issue was fixed by adding a variable named "hasCollector" which allows the program to only enable the config button when the euRun has a datacollector.
Additionally this commit fixes an issue with the Run where after a run one had
to reconfigure the connections. As of now, the program relies on a varible :
needsConfigureCheck to determine when it is appropriate to check the configure
status of the connections. This variable is set to false after a run and set to
true once a connection is added.