Fixing occasionally failing End Of Run (EOR)
Due to a wrong sequence of commands and transaction with the DAQ board,
there was the possibility of a ongoing read transaction while the DAQ was stopped,
which caused the DAQ being stuck during EOR. This has been fixed at least for the
event-based readout mode. The packet-based readout mode is to be verified.