Changes to be committed: 	modified:   main/include/eudaq/CommandReceiver.hh 	modified:   main/include/eudaq/Status.hh 	modified:   main/lib/src/CommandReceiver.cc 	modified:   main/lib/src/Status.cc
 Untracked files:
	main/include/eudaq/MachineState.hh
	main/lib/src/MachineState.cc

Addition of a MachineState state to the status class. This includes an enum for the states, and
function to associate the int to the string. Getter for the status added. State added to the
serializer.

In CommandReceiver setStatus was modified so that one can pass the machine state at the same
time as other variables