On branch fsm_beryl  Changes to be committed: 	modified:   gui/include/RunControlModel.hh 	modified:   gui/include/euLog.hh 	modified:   main/exe/src/ExampleProducer.cxx 	modified:   main/exe/src/RunListener.cxx 	modified:   main/exe/src/TestDataCollector.cxx 	modified:   main/exe/src/TestLogCollector.cxx 	modified:   main/exe/src/TestMonitor.cxx 	modified:   main/exe/src/TestProducer.cxx 	modified:   main/include/eudaq/CommandReceiver.hh 	modified:   main/include/eudaq/LogCollector.hh 	modified:   main/include/eudaq/Logger.hh 	modified:   main/include/eudaq/RunControl.hh 	modified:   main/include/eudaq/TransportBase.hh 	modified:   main/include/eudaq/TransportTCP.hh 	modified:   main/lib/src/CommandReceiver.cc 	modified:   main/lib/src/DataCollector.cc 	modified:   main/lib/src/Monitor.cc 	modified:   main/lib/src/RunControl.cc 	modified:   main/lib/src/Status.cc 	modified:   main/lib/src/TransportBase.cc 	modified:   main/python/src/PyDataCollector.cpp 	modified:   main/python/src/PyLogCollector.cpp 	modified:   main/python/src/PyProducer.cpp 	modified:   monitors/onlinemon/include/OnlineMon.hh 	modified:   monitors/onlinemon/src/OnlineMon.cxx
 Untracked files:
	main/include/eudaq/MachineState.hh
	main/lib/src/MachineState.cc

The example producers have to been modified to emmit States as defined within
the Status class. A Class called Machine State has been defined to handle these
states. One has been declared within the RunControl. This MachineState tracks
the states of all the current connection and allows for easy access to a state
for the entire system.

The ConnectionInfo and ConnectionInfoTCP have both been modified to accomdate
the map defined in Machine state. The operator < has been defined in
ConnectionInfo and inhereted in ConnectionInfoTCP. Additional the variable
m_host has been moved from the ConnectionInfoTCP class to the ConnectionInfo
class, once again so that the maps may be logically ordered.

The < operator that is defined in ConnectionInfo aquires the address of the
connection and does a int comparison with that of another ConnectionInfo.

Currently the RunControl prints out via cout the current state of the machine
and all the connections every time those connections are updated. Nothing
relies on the state of the machine being known.